Refactor: Response 클래스 생성 #87
